Now working in terms of 16 bit shorts instead of bytes making it so we need signed values.

+++ emulator/glue/Core.cs
@@ -214,7 +214,7 @@ ioRegisters,
         CPU.Step();
     }
 
-    public (byte left, byte right) Sample() => APU.Sample();
+    public (short left, short right) Sample() => APU.Sample();
 
     protected virtual void Dispose(bool disposing)
     {
